+++ tests/utils/clearAllAnnotations.ts
@@ -1,0 +1,10 @@
+const clearAllAnnotations = async page => {
+  await page.evaluate(
+    ({ cornerstoneTools }: AppTypes.Test) => {
+      cornerstoneTools.annotation.state.removeAllAnnotations();
+    },
+    await page.evaluateHandle('window')
+  );
+};
+
+export { clearAllAnnotations };
