+++ tests/utils/index.ts
@@ -4,6 +4,19 @@ import { screenShotPaths } from './screenShotPaths';
 import { simulateClicksOnElement } from './simulateClicksOnElement';
 import { reduce3DViewportSize } from './reduce3DviewportSize';
 import { getMousePosition, initilizeMousePositionTracker } from './mouseUtils';
+import { getSUV } from './getSUV';
+import { getTMTVModalityUnit } from './getTMTVModalityUnit';
+import { clearAllAnnotations } from './clearAllAnnotations';
 
-
-export { visitStudy, checkForScreenshot, screenShotPaths, simulateClicksOnElement, reduce3DViewportSize, getMousePosition, initilizeMousePositionTracker };
+export {
+  visitStudy,
+  checkForScreenshot,
+  screenShotPaths,
+  simulateClicksOnElement,
+  reduce3DViewportSize,
+  getMousePosition,
+  initilizeMousePositionTracker,
+  getSUV,
+  getTMTVModalityUnit,
+  clearAllAnnotations,
+};
